move top-level "util" directory under "src" also use console.log instead of trace when missing "updatedAt" since the trace doesn't give us anything useful since we just have the stack in this repo
