+++ src/util/pubSubStats.js
@@ -40,7 +40,7 @@ function trackSubscribe(evt, obj) {
 
   /*
    * Calculate the elapsed time. If we can't find an "updatedAt" attribute,
-   * treat the elapsed time as 0 but console.trace the object.
+   * treat the elapsed time as 0 but log the object.
    */
   let elapsed = 0;
   let updatedAtFromObj;
@@ -54,7 +54,7 @@ function trackSubscribe(evt, obj) {
     nameFromObj = obj.new.name;
     elapsed = now - new Date(obj.new.updatedAt);
   } else {
-    console.trace('Where is updatedAt? ' + JSON.stringify(obj));
+    console.log('Missing updatedAt: ' + JSON.stringify(obj));
   }
 
   if (elapsed > 2000) {
